Reformatted the code. client_details.py -> Now handles the client details Class processor_base.py -> All the base classes for the processors processors.py -> All available processors registry.py -> Registry mechanism for the processors
